Make sure permissions are granted on tables post-migration.
